requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
coverage==7.13.2
cryptography==42.0.8
Django==6.0.1
django-extensions==4.1
django-allauth==65.14.2
idna==3.11
iniconfig==2.3.0
packaging==26.0
pillow==12.1.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-mock==3.15.1
python-dotenv==1.2.1
pywhatkit==5.4
reportlab==4.4.10
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3

# Production dependencies
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
djangorestframework==3.14.0
